api/requirements.txt
flask
pymongo
python-dotenv
bcrypt
PyJWT
